requirements.txt
attrs==26.1.0
numpy==2.4.4
pillow==12.1.1
pyarrow==23.0.1
rerun-sdk==0.31.0
typing_extensions==4.15.0
